Back out rpc/send return type specializations from codegen.
As Matthias points out, the exceptions from the pattern in AMQP make
those awkward to deal with and just move one annoyance to another
place.  Thus, we revert to the simpler structure of putting rpc
(synchronous) and call (asynchronous) Java methods on the interface
and impl for Method, so that method transmission can be easily chained
on to Builder-based constructions of Methods.

--HG--
branch : bug23451